els-ASAP auto commit file/tag lldpad-0.9.46-9.el7_0.src.rpm

+++ src/test/nltest.c
@@ -1156,14 +1156,12 @@ void print_pfc(struct ieee_pfc *pfc)
 
 	printf("\t requests: ");
 	for (i = 0; i < 8; i++)
-		printf("%llu ",
-		       (unsigned long long)pfc->requests[i]);
+		printf("%lli ", pfc->requests[i]);
 	printf("\n");
 
 	printf("\t requests: ");
 	for (i = 0; i < 8; i++)
-		printf("%llu ",
-		       (unsigned long long)pfc->indications[i]);
+		printf("%lli ", pfc->indications[i]);
 	printf("\n");
 }
